mirage: Split route handlers into multiple files
